+++ resources/hpxml-measures/HPXMLtoOpenStudio/tests/test_miscloads.rb
@@ -8,13 +8,17 @@ require_relative '../measure.rb'
 require_relative '../resources/util.rb'
 
 class HPXMLtoOpenStudioMiscLoadsTest < Minitest::Test
-  def teardown
-    File.delete(File.join(File.dirname(__FILE__), 'results_annual.csv')) if File.exist? File.join(File.dirname(__FILE__), 'results_annual.csv')
-    File.delete(File.join(File.dirname(__FILE__), 'results_design_load_details.csv')) if File.exist? File.join(File.dirname(__FILE__), 'results_design_load_details.csv')
+  def setup
+    @root_path = File.absolute_path(File.join(File.dirname(__FILE__), '..', '..'))
+    @sample_files_path = File.join(@root_path, 'workflow', 'sample_files')
+    @tmp_hpxml_path = File.join(@sample_files_path, 'tmp.xml')
   end
 
-  def sample_files_dir
-    return File.join(File.dirname(__FILE__), '..', '..', 'workflow', 'sample_files')
+  def teardown
+    File.delete(@tmp_hpxml_path) if File.exist? @tmp_hpxml_path
+    File.delete(File.join(File.dirname(__FILE__), 'in.schedules.csv')) if File.exist? File.join(File.dirname(__FILE__), 'in.schedules.csv')
+    File.delete(File.join(File.dirname(__FILE__), 'results_annual.csv')) if File.exist? File.join(File.dirname(__FILE__), 'results_annual.csv')
+    File.delete(File.join(File.dirname(__FILE__), 'results_design_load_details.csv')) if File.exist? File.join(File.dirname(__FILE__), 'results_design_load_details.csv')
   end
 
   def get_kwh_therm_per_year(model, name)
@@ -43,7 +47,7 @@ class HPXMLtoOpenStudioMiscLoadsTest < Minitest::Test
 
   def test_misc_loads
     args_hash = {}
-    args_hash['hpxml_path'] = File.absolute_path(File.join(sample_files_dir, 'base.xml'))
+    args_hash['hpxml_path'] = File.absolute_path(File.join(@sample_files_path, 'base.xml'))
     model, _hpxml, _hpxml_bldg = _test_measure(args_hash)
 
     # Check misc plug loads
@@ -75,7 +79,7 @@ class HPXMLtoOpenStudioMiscLoadsTest < Minitest::Test
 
   def test_large_uncommon_loads
     args_hash = {}
-    args_hash['hpxml_path'] = File.absolute_path(File.join(sample_files_dir, 'base-misc-loads-large-uncommon.xml'))
+    args_hash['hpxml_path'] = File.absolute_path(File.join(@sample_files_path, 'base-misc-loads-large-uncommon.xml'))
     model, _hpxml, _hpxml_bldg = _test_measure(args_hash)
 
     # Check misc plug loads
@@ -136,7 +140,7 @@ class HPXMLtoOpenStudioMiscLoadsTest < Minitest::Test
 
   def test_large_uncommon_loads2
     args_hash = {}
-    args_hash['hpxml_path'] = File.absolute_path(File.join(sample_files_dir, 'base-misc-loads-large-uncommon2.xml'))
+    args_hash['hpxml_path'] = File.absolute_path(File.join(@sample_files_path, 'base-misc-loads-large-uncommon2.xml'))
     model, _hpxml, _hpxml_bldg = _test_measure(args_hash)
 
     # Check misc plug loads
@@ -195,9 +199,98 @@ class HPXMLtoOpenStudioMiscLoadsTest < Minitest::Test
     assert_in_delta(55, therm_yr, 1.0)
   end
 
-  def test_operational_defaults
+  def test_operational_0_occupants
     args_hash = {}
-    args_hash['hpxml_path'] = File.absolute_path(File.join(sample_files_dir, 'base-residents-5-5.xml'))
+    args_hash['hpxml_path'] = @tmp_hpxml_path
+    hpxml, hpxml_bldg = _create_hpxml('base-residents-0.xml')
+    [HPXML::PlugLoadTypeElectricVehicleCharging,
+     HPXML::PlugLoadTypeWellPump].each do |plug_load_type|
+      hpxml_bldg.plug_loads.add(id: "PlugLoad#{hpxml_bldg.plug_loads.size + 1}",
+                                plug_load_type: plug_load_type)
+    end
+    [HPXML::FuelLoadTypeFireplace,
+     HPXML::FuelLoadTypeGrill,
+     HPXML::FuelLoadTypeLighting].each do |fuel_load_type|
+      hpxml_bldg.fuel_loads.add(id: "FuelLoad#{hpxml_bldg.fuel_loads.size + 1}",
+                                fuel_type: HPXML::FuelTypeNaturalGas,
+                                fuel_load_type: fuel_load_type,
+                                therm_per_year: 100)
+    end
+    hpxml_bldg.pools.add(id: "Pool#{hpxml_bldg.pools.size + 1}",
+                         type: HPXML::TypeUnknown,
+                         pump_type: HPXML::TypeUnknown,
+                         heater_type: HPXML::HeaterTypeGas,
+                         heater_load_units: HPXML::UnitsThermPerYear)
+    hpxml_bldg.permanent_spas.add(id: "PermanentSpa#{hpxml_bldg.permanent_spas.size + 1}",
+                                  type: HPXML::TypeUnknown,
+                                  pump_type: HPXML::TypeUnknown,
+                                  pump_kwh_per_year: 100,
+                                  heater_type: HPXML::HeaterTypeElectricResistance,
+                                  heater_load_units: HPXML::UnitsKwhPerYear,
+                                  heater_load_value: 100)
+
+    XMLHelper.write_file(hpxml.to_doc, @tmp_hpxml_path)
+    model, _hpxml, _hpxml_bldg = _test_measure(args_hash)
+
+    # Check misc plug loads
+    kwh_yr, therm_yr = get_kwh_therm_per_year(model, Constants::ObjectTypeMiscPlugLoads)
+    assert_equal(0, kwh_yr)
+    assert_equal(0, therm_yr)
+
+    # Check television
+    kwh_yr, therm_yr = get_kwh_therm_per_year(model, Constants::ObjectTypeMiscTelevision)
+    assert_equal(0, kwh_yr)
+    assert_equal(0, therm_yr)
+
+    # Check vehicle
+    kwh_yr, therm_yr = get_kwh_therm_per_year(model, Constants::ObjectTypeMiscElectricVehicleCharging)
+    assert_equal(0, kwh_yr)
+    assert_equal(0, therm_yr)
+
+    # Check well pump
+    kwh_yr, therm_yr = get_kwh_therm_per_year(model, Constants::ObjectTypeMiscWellPump)
+    assert_equal(0, kwh_yr)
+    assert_equal(0, therm_yr)
+
+    # Check pool pump
+    kwh_yr, therm_yr = get_kwh_therm_per_year(model, Constants::ObjectTypeMiscPoolPump)
+    refute_equal(0, kwh_yr)
+    assert_equal(0, therm_yr)
+
+    # Check pool heater
+    kwh_yr, therm_yr = get_kwh_therm_per_year(model, Constants::ObjectTypeMiscPoolHeater)
+    assert_equal(0, kwh_yr)
+    refute_equal(0, therm_yr)
+
+    # Check permanent spa pump
+    kwh_yr, therm_yr = get_kwh_therm_per_year(model, Constants::ObjectTypeMiscPermanentSpaPump)
+    refute_equal(0, kwh_yr)
+    assert_equal(0, therm_yr)
+
+    # Check permanent spa heater
+    kwh_yr, therm_yr = get_kwh_therm_per_year(model, Constants::ObjectTypeMiscPermanentSpaHeater)
+    refute_equal(0, kwh_yr)
+    assert_equal(0, therm_yr)
+
+    # Check grill
+    kwh_yr, therm_yr = get_kwh_therm_per_year(model, Constants::ObjectTypeMiscGrill)
+    assert_equal(0, kwh_yr)
+    assert_equal(0, therm_yr)
+
+    # Check lighting
+    kwh_yr, therm_yr = get_kwh_therm_per_year(model, Constants::ObjectTypeMiscLighting)
+    assert_equal(0, kwh_yr)
+    assert_equal(0, therm_yr)
+
+    # Check fireplace
+    kwh_yr, therm_yr = get_kwh_therm_per_year(model, Constants::ObjectTypeMiscFireplace)
+    assert_equal(0, kwh_yr)
+    assert_equal(0, therm_yr)
+  end
+
+  def test_operational_5_5_occupants
+    args_hash = {}
+    args_hash['hpxml_path'] = File.absolute_path(File.join(@sample_files_path, 'base-residents-5-5.xml'))
     model, _hpxml, _hpxml_bldg = _test_measure(args_hash)
 
     # Check misc plug loads
@@ -258,7 +351,7 @@ class HPXMLtoOpenStudioMiscLoadsTest < Minitest::Test
 
   def test_operational_large_uncommon_loads
     args_hash = {}
-    args_hash['hpxml_path'] = File.absolute_path(File.join(sample_files_dir, 'base-residents-1-misc-loads-large-uncommon.xml'))
+    args_hash['hpxml_path'] = File.absolute_path(File.join(@sample_files_path, 'base-residents-1-misc-loads-large-uncommon.xml'))
     model, _hpxml, _hpxml_bldg = _test_measure(args_hash)
 
     # Check misc plug loads
@@ -319,7 +412,7 @@ class HPXMLtoOpenStudioMiscLoadsTest < Minitest::Test
 
   def test_operational_large_uncommon_loads2
     args_hash = {}
-    args_hash['hpxml_path'] = File.absolute_path(File.join(sample_files_dir, 'base-residents-1-misc-loads-large-uncommon2.xml'))
+    args_hash['hpxml_path'] = File.absolute_path(File.join(@sample_files_path, 'base-residents-1-misc-loads-large-uncommon2.xml'))
     model, _hpxml, _hpxml_bldg = _test_measure(args_hash)
 
     # Check misc plug loads
@@ -414,5 +507,10 @@ class HPXMLtoOpenStudioMiscLoadsTest < Minitest::Test
     File.delete(File.join(File.dirname(__FILE__), 'in.xml'))
 
     return model, hpxml, hpxml.buildings[0]
+  end
+
+  def _create_hpxml(hpxml_name)
+    hpxml = HPXML.new(hpxml_path: File.join(@sample_files_path, hpxml_name))
+    return hpxml, hpxml.buildings[0]
   end
 end