+++ resources/hpxml-measures/tasks.rb
@@ -2624,6 +2624,14 @@ if ARGV[0].to_sym == :update_measures
   puts 'Applying rubocop auto-correct to measures...'
   system(command)
 
+  # Update a BuildResidentialHPXML/resources file when the OS-HPXML version changes.
+  # This will ensure that the BuildResidentialHPXML measure.xml is appropriately updated.
+  # Without this, the BuildResidentialHPXML measure has no differences and so OpenStudio
+  # would skip updating it.
+  version_rb_path = File.join(File.dirname(__FILE__), 'HPXMLtoOpenStudio/resources/version.rb')
+  version_txt_path = File.join(File.dirname(__FILE__), 'BuildResidentialHPXML/resources/version.txt')
+  File.write(version_txt_path, Digest::MD5.file(version_rb_path).hexdigest)
+
   # Update measures XMLs
   puts 'Updating measure.xmls...'
   Dir['**/measure.xml'].each do |measure_xml|